fixed compile errors

+++ src/se/crafted/chrisb/ecoCreature/entities/ecoBlockListener.java
@@ -33,7 +33,7 @@ public class ecoBlockListener extends BlockListener
           k = j;
         else
           k = localRandom.nextInt(j) + 1;
-        ItemStack localItemStack = new ItemStack((int)ecoConstants.CD[14][i][0], k, 0);
+        ItemStack localItemStack = new ItemStack((int)ecoConstants.CD[14][i][0], k, (short)0);
         if (localItemStack == null)
           continue;
         if (localItemStack.getAmount() == 0)
